janiecee CPPLYNT

+++ src/s21_string.h
@@ -8,7 +8,7 @@
 #define _S21_STRING_H_
 
 #define s21_NULL ((void*)0)
-typedef unsigned s21_size_t;
+typedef long unsigned s21_size_t;
 
 void *s21_memcpy(void *dest, const void *src, s21_size_t n);
 void *s21_memset(void *str, int c, s21_size_t n);
@@ -18,5 +18,6 @@ char *s21_strncpy(char *dest, const char *src, s21_size_t n);
 s21_size_t s21_strcspn(const char *str1, const char *str2);
 char *s21_strerror(int errnum);
 char *s21_strrchr(const char *str, int c);
+// int sprintf(char *str, const char *format, ...);
 
 #endif
